+++ .github/workflows/build.yaml
@@ -112,7 +112,7 @@ jobs:
       # AWS ECR Public Registry related steps
       - name: Import AWS Secrets
         uses: hashicorp/vault-action@v2.7.3
-        if: ${{ startsWith(github.ref, 'refs/tags') }}
+#        if: ${{ startsWith(github.ref, 'refs/tags') }}
         with:
           url: https://hashicorp-vault.awsp.sneaksanddata.com/
           role: github
@@ -124,20 +124,20 @@ jobs:
 
       - name: Configure AWS Credentials
         uses: aws-actions/configure-aws-credentials@v4.0.0
-        if: ${{ startsWith(github.ref, 'refs/tags') }}
+#        if: ${{ startsWith(github.ref, 'refs/tags') }}
         with:
           aws-access-key-id: ${{ env.ACCESS_KEY_ID }}
           aws-secret-access-key: ${{ env.ACCESS_KEY }}
           aws-region: us-east-1
 
       - name: Login to Amazon ECR Public
-        if: ${{ startsWith(github.ref, 'refs/tags') }}
+#        if: ${{ startsWith(github.ref, 'refs/tags') }}
         uses: aws-actions/amazon-ecr-login@v2
         with:
           registry-type: public
           
       - name: Push image to ECR Public registry
-        if: ${{ startsWith(github.ref, 'refs/tags') }}
+#        if: ${{ startsWith(github.ref, 'refs/tags') }}
         uses: akhilerm/tag-push-action@v2.1.0
         with:
           src: ${{ steps.meta.outputs.tags }}
