+++ src/Models/RestApiFixedAuthStreamContext.cs
@@ -52,11 +52,6 @@ public class RestApiFixedAuthStreamContext : RestApiFixedAuthBase
     public TimeSpan GroupingInterval { get; init; }
 
     /// <summary>
-    /// Data location for parquet files.
-    /// </summary>
-    public string SinkLocation { get; init; }
-
-    /// <summary>
     /// Number of seconds to look back when determining first set of changes to extract.
     /// </summary>
     public int LookbackInterval { get; init; }
